requirements.txt
# Discord Bot Requirements
discord.py>=2.3.0
aiohttp>=3.9.0

# Optional but recommended for better performance
uvloop>=0.19.0

# For async HTTP requests (used in PostManager)
# aiohttp is already included above
